Updated the tocantera.sh script.
Accidentally committed some of the changes to this file as
part of earlier commits, but here's the commit message that should
have gone with it :-)

It now can be used in SLURM, or without.

$ sbatch tocantera.sh

will run an array job, with 20 simultaneous tasks,
trying to convert the first 520 chemkin files into
cantera files.

If you just run

$ ./tocantera.sh

then it will just do the chem-annotated version.